Move `component` subcommands to `cli/component.py`
